+++ server.test.js
@@ -7,10 +7,11 @@ const timeout = AbortSignal.timeout(60_000);
 describe('Test HTTP server', { concurrency: 5, signal: timeout }, async () => {
 	let port = 8000;
 	const staticRoot = '/static/';
+	const controller = new AbortController();
 	const routes = {
 		'/tasks': '@shgysk8zer0/http-server/api/tasks.js',
+		'/': '@shgysk8zer0/http-server/api/home.js',
 	};
-	const controller = new AbortController();
 
 	test('Test basic, static requests', { signal: timeout }, async () => {
 		try {
@@ -81,6 +82,66 @@ describe('Test HTTP server', { concurrency: 5, signal: timeout }, async () => {
 			const resp = await fetch(tasksURL, { signal });
 			const tasks = await resp.json();
 			deepStrictEqual(tasks, [{ id, created, completed, ...task }], 'Dynamic routes should work.');
+			server.close();
+		} catch(err) {
+			controller.abort(err);
+			fail(err);
+		}
+	});
+
+	test('Check pre/post processors.', { signal: timeout }, async () => {
+		try {
+			const signal = AbortSignal.any([timeout, controller.signal]);
+
+			const { url, server } = await serve({
+				port: port++,
+				signal,
+				staticRoot,
+				logger: null,
+				routes,
+				requestPreprocessors: [
+					(req, { reject }) => {
+						if (req.headers.has('X-Foo')) {
+							reject(new Error('That is not allowed.'));
+						}
+					}
+				],
+				responsePostprocessors: [
+					resp => resp.headers.set('Access-Control-Allow-Origin', '*'),
+				]
+			});
+
+			const allowed = await fetch(url);
+			const blocked = await fetch(url, { headers: { 'X-Foo': 'bar' }});
+			ok(allowed.ok, '`requestPreprocessors` should allow valid requests.');
+			ok(! blocked.ok, '`requestPreprocessors` should not allow invalid requests.');
+			strictEqual(allowed.headers.get('Access-Control-Allow-Origin'), '*', '`responsePostprocessors` should add CORS headers to allowed requests.');
+			strictEqual(blocked.headers.get('Access-Control-Allow-Origin'), '*', '`responsePostprocessors` should add CORS headers to disallowed requests.');
+			server.close();
+		} catch(err) {
+			controller.abort(err);
+			fail(err);
+		}
+	});
+
+	test('Verify that server does not crash if signal aborts.', { signal: timeout }, async () => {
+		try {
+			const { url, server } = await serve({
+				port: port++,
+				signal: controller.signal,
+				staticRoot,
+				logger: null,
+				routes,
+				requestPreprocessors: [
+					(req, { controller }) => {
+						controller.abort('Just testing...');
+					}
+				]
+			});
+
+			const resp = await fetch(url, { signal: controller.signal });
+			strictEqual(resp.status, 408, 'Aborted requests should have a 4xx status code.');
+			strictEqual(resp.headers.get('Content-Type'), 'application/json', 'Aborted requests should should respond with an error as JSON.');
 			server.close();
 		} catch(err) {
 			controller.abort(err);